+++ src/main/java/com/example/myspringapp/resourse/UserResourse.java
@@ -1,13 +1,19 @@
 package com.example.myspringapp.resourse;
 
 
+import com.example.myspringapp.exception.InvalidTokenException;
 import com.example.myspringapp.exception.NameNotAllowedException;
+import com.example.myspringapp.exception.UserNotAuthorizedException;
+import com.example.myspringapp.model.FirebaseUser;
 import com.example.myspringapp.model.User;
+import com.example.myspringapp.service.FirebaseService;
 import com.example.myspringapp.service.UserService;
+import com.google.firebase.auth.FirebaseAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.List;
 
 
@@ -18,16 +24,35 @@ public class UserResourse {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private FirebaseService firebaseService;
+
     @PostMapping
-    public User saveUser(@RequestBody @Valid  User user){
-        return userService.saveUser(user);
+    public User saveUser(@RequestBody @Valid  User user, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException, NameNotAllowedException {
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if(firebaseUser != null){
+            if (user.getName().equalsIgnoreCase("root")) {
+                throw new NameNotAllowedException();
+            }
+            user.setEmail(firebaseUser.getEmail());
+            return userService.saveUser(user);
+        } else {
+            throw new IOException();
+        }
+
     }
 
-    @GetMapping
-    public List<User> getAllUsers(){
-        return userService.getAllUsers();
-    }
+    @GetMapping("/all")
+    public List<User> getAllUsers(@RequestHeader(name = "idToken") String idToken)
+            throws FirebaseAuthException,
+             IOException {
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if (firebaseUser != null) {
+            return userService.getAllUsers();
+        }
 
+        return null;
+    }
     @GetMapping("/find-by-name")
     public List<User> getUserByName(@RequestParam(name = "name") String name) throws NameNotAllowedException {
         if(name.equalsIgnoreCase("root")){
@@ -38,13 +63,58 @@ public class UserResourse {
 
 
     @PutMapping
-    public User updateUser(@RequestBody User user){
-        return userService.updateUser(user);
+    public User updateUser(@RequestBody @Valid User user,
+                           @RequestHeader(name = "idToken") String idToken)
+            throws NameNotAllowedException,
+            IOException,
+            FirebaseAuthException,
+            InvalidTokenException,
+            UserNotAuthorizedException {
+
+        if (user.getName().equalsIgnoreCase("root")) {
+            throw new NameNotAllowedException();
+        }
+
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+
+        if (firebaseUser != null) {
+            if (!firebaseUser.getEmail().equalsIgnoreCase(user.getEmail())) {
+                throw new UserNotAuthorizedException(
+                        "FireBaseUser.email: " + firebaseUser.getEmail() + "\n" +
+                                "User.email: " + user.getEmail()
+                );
+            } else {
+                return userService.updateUser(user);
+            }
+        } else {
+            throw new InvalidTokenException();
+        }
     }
 
     @DeleteMapping
-    public void deleteUser(@RequestParam(name = "userId") String userId){
-        userService.deleteUser(userId);
+    public String deleteUser(@RequestParam(name = "userId") String userId,
+                             @RequestHeader(name = "idToken") String idToken)
+            throws IOException,
+            FirebaseAuthException,
+            InvalidTokenException,
+            UserNotAuthorizedException {
+
+        User user = userService.getUserById(userId);
+
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+
+        if (firebaseUser != null) {
+            if (!firebaseUser.getEmail().equalsIgnoreCase(user.getEmail())) {
+                throw new UserNotAuthorizedException(
+                        "FireBaseUser.email: " + firebaseUser.getEmail() + "\n" +
+                                "User.email: " + user.getEmail()
+                );
+            } else {
+                return userService.deleteUser(userId);
+            }
+        } else {
+            throw new InvalidTokenException();
+        }
     }
 
 }
